REF: put contrast_allpairs directly in contrast, avoiding sandbox imports
