+++ statsmodels/sandbox/stats/contrast_tools.py
@@ -26,29 +26,10 @@ from __future__ import print_function
 from statsmodels.compat.python import zip
 import numpy as np
 
-#next 3 functions copied from multicomp.py
+from statsmodels.stats.contrast import contrast_allpairs
 
-def contrast_allpairs(nm):
-    '''contrast or restriction matrix for all pairs of nm variables
 
-    Parameters
-    ----------
-    nm : int
-
-    Returns
-    -------
-    contr : ndarray, 2d, (nm*(nm-1)/2, nm)
-       contrast matrix for all pairwise comparisons
-
-    '''
-    contr = []
-    for i in range(nm):
-        for j in range(i+1, nm):
-            contr_row = np.zeros(nm)
-            contr_row[i] = 1
-            contr_row[j] = -1
-            contr.append(contr_row)
-    return np.array(contr)
+# next 2 functions moved from multicomp.py
 
 def contrast_all_one(nm):
     '''contrast or restriction matrix for all against first comparison
