+++ statsmodels/sandbox/stats/multicomp.py
@@ -90,6 +90,7 @@ from statsmodels.iolib.table import SimpleTable
 from numpy.testing import assert_almost_equal, assert_equal
 #temporary circular import
 from statsmodels.stats.multitest import multipletests, _ecdf as ecdf, fdrcorrection as fdrcorrection0, fdrcorrection_twostage
+from statsmodels.stats.contrast import contrast_allpairs
 from statsmodels.graphics import utils
 from statsmodels.tools.sm_exceptions import ValueWarning
 
@@ -1439,59 +1440,6 @@ def distance_st_range(mean_all, nobs_all, var_all, df=None, triu=False):
 
     return st_range, meandiffs, std_pairs, (idx1,idx2)  #return square arrays
 
-
-def contrast_allpairs(nm):
-    '''contrast or restriction matrix for all pairs of nm variables
-
-    Parameters
-    ----------
-    nm : int
-
-    Returns
-    -------
-    contr : ndarray, 2d, (nm*(nm-1)/2, nm)
-       contrast matrix for all pairwise comparisons
-
-    '''
-    contr = []
-    for i in range(nm):
-        for j in range(i+1, nm):
-            contr_row = np.zeros(nm)
-            contr_row[i] = 1
-            contr_row[j] = -1
-            contr.append(contr_row)
-    return np.array(contr)
-
-def contrast_all_one(nm):
-    '''contrast or restriction matrix for all against first comparison
-
-    Parameters
-    ----------
-    nm : int
-
-    Returns
-    -------
-    contr : ndarray, 2d, (nm-1, nm)
-       contrast matrix for all against first comparisons
-
-    '''
-    contr = np.column_stack((np.ones(nm-1), -np.eye(nm-1)))
-    return contr
-
-def contrast_diff_mean(nm):
-    '''contrast or restriction matrix for all against mean comparison
-
-    Parameters
-    ----------
-    nm : int
-
-    Returns
-    -------
-    contr : ndarray, 2d, (nm-1, nm)
-       contrast matrix for all against mean comparisons
-
-    '''
-    return np.eye(nm) - np.ones((nm,nm))/nm
 
 def tukey_pvalues(std_range, nm, df):
     #corrected but very slow with warnings about integration
